requirements.txt
python-dotenv~=1.0.1
prettytable~=3.9.0
paramiko~=3.4.0